refactor(documents): Collection.clientId 타입을 Long → Integer 로 정합화
DDL 은 client_id INT 이고 다른 엔티티(ProformaInvoice/PurchaseOrder 등)는
Integer 로 선언돼 있는데 Collection 만 Long 으로 드리프트돼 있었다.
Hibernate 자동 변환으로 런타임 문제는 없었지만 DTO/Query/Response 전수에
Long 이 흘러나와 일관성 해침.

[변경]
- Collection.clientId: Long → Integer
- Collection 생성자 파라미터 Long → Integer (2곳)
- Collection.getClientId() 반환 Integer
- CollectionView.clientId: Long → Integer
- DocumentCommandController.CollectionResponse.clientId: Long → Integer
- DocumentQueryController.CollectionResponse.clientId: Long → Integer
- PurchaseOrderCreationService.createInitialViews: .longValue() 제거

Batch 3 (DB 타입 정합성) 의 clientId 부분 해소. piId 는 이미 entity 기준
VARCHAR 로 DDL 맞춤 완료 (root@9a03c52).

+++ src/main/java/com/team2/documents/command/application/controller/DocumentCommandController.java
@@ -725,7 +725,7 @@ public class DocumentCommandController {
             @Schema(description = "수금 ID") Long collectionId,
             @Schema(description = "PO 문서 ID") String poId,
             @Schema(description = "PO 문서 번호") String poNo,
-            @Schema(description = "거래처 ID") Long clientId,
+            @Schema(description = "거래처 ID") Integer clientId,
             @Schema(description = "거래처명") String clientName,
             @Schema(description = "총 금액") BigDecimal totalAmount,
             @Schema(description = "수금 완료 금액") BigDecimal collectedAmount,
